+++ sandbox/auth-ms-basic-example/src/application.ts
@@ -16,6 +16,7 @@ import {
   SignUpBindings,
 } from '@sourceloop/authentication-service';
 import path from 'path';
+import {AzureAdSignupProvider} from './providers';
 import {LocalSignupProvider} from './providers/local-signup.provider';
 import {MySequence} from './sequence';
 
@@ -41,6 +42,9 @@ export class AuthMultitenantExampleApplication extends BootMixin(
     this.component(AuthenticationServiceComponent);
     this.bind(SignUpBindings.LOCAL_SIGNUP_PROVIDER).toProvider(
       LocalSignupProvider,
+    );
+    this.bind(SignUpBindings.AZURE_AD_SIGN_UP_PROVIDER).toProvider(
+      AzureAdSignupProvider,
     );
 
     this.projectRoot = __dirname;