+++ services/chat-service/src/models/index.ts
@@ -4,3 +4,5 @@
 // https://opensource.org/licenses/MIT
 export * from './message.model';
 export * from './message-recipient.model';
+export * from './attachment-file-dto.model';
+export * from './attachment-file.model';