+++ sandbox/chat-notification-pubnub-example/services/notifications-service/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.3.11 (2022-10-10)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example_notif-service
+
+
+
+
+
+## <small>0.3.10 (2022-09-26)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example_notif-service
+
+
+
+
+
+## <small>0.3.9 (2022-09-21)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example_notif-service
+
+
+
+
+
+## <small>0.3.8 (2022-09-21)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example_notif-service
+
+
+
+
+
+## <small>0.3.7 (2022-09-20)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example_notif-service
+
+
+
+
+
 ## <small>0.3.6 (2022-09-12)</small>
 
 **Note:** Version bump only for package @sourceloop/chat-notif-example_notif-service