+++ sandbox/chat-notification-socketio-example/facade/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.1.8 (2022-10-10)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example-facade-socket
+
+
+
+
+
+## <small>0.1.7 (2022-09-26)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example-facade-socket
+
+
+
+
+
+## <small>0.1.6 (2022-09-21)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example-facade-socket
+
+
+
+
+
+## <small>0.1.5 (2022-09-20)</small>
+
+**Note:** Version bump only for package @sourceloop/chat-notif-example-facade-socket
+
+
+
+
+
 ## <small>0.1.4 (2022-09-12)</small>
 
 **Note:** Version bump only for package @sourceloop/chat-notif-example-facade-socket