+++ sandbox/feature-toggle-example/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## <small>0.5.11 (2022-10-10)</small>
+
+**Note:** Version bump only for package @sourceloop/feature-toggle-example
+
+
+
+
+
+## <small>0.5.10 (2022-09-26)</small>
+
+**Note:** Version bump only for package @sourceloop/feature-toggle-example
+
+
+
+
+
+## <small>0.5.9 (2022-09-21)</small>
+
+**Note:** Version bump only for package @sourceloop/feature-toggle-example
+
+
+
+
+
+## <small>0.5.8 (2022-09-21)</small>
+
+**Note:** Version bump only for package @sourceloop/feature-toggle-example
+
+
+
+
+
+## <small>0.5.7 (2022-09-20)</small>
+
+**Note:** Version bump only for package @sourceloop/feature-toggle-example
+
+
+
+
+
 ## <small>0.5.6 (2022-09-12)</small>
 
 **Note:** Version bump only for package @sourceloop/feature-toggle-example